+++ server/src/middleware/guards/auth.ts
@@ -1,6 +1,6 @@
 import { Context } from 'elysia';
 import { BaseResultData } from '@/core/result';
-import { VerifyToken, ParseToken } from '@/shared/jwt';
+import { VerifyToken } from '@/shared/jwt';
 import { CacheEnum } from '@/constants/enum';
 import { Get } from '@/core/database/redis';
 
@@ -16,9 +16,7 @@ export async function AuthGuard(ctx: Context) {
         if (!auth) return BaseResultData.fail(401);
         const token = auth.split(' ')[1];
         if (!token) return BaseResultData.fail(401);
-        const isVerify = await VerifyToken('accessToken', token);
-        if (!isVerify) return BaseResultData.fail(401);
-        const payload = await ParseToken(token);
+        const payload = await VerifyToken('accessToken', token);
         if (!payload?.userId) return BaseResultData.fail(401);
         const user = await Get(CacheEnum.ONLINE_USER + payload.userId);
         if (!user) return BaseResultData.fail(401);